requirements-fastapi.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
asyncpg>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyJWT>=2.10.0
bcrypt>=4.0.0
